make tests work
